refactor(skeleton): per-component derivative modules with own docs & stories
Reorganize the Skeleton family so each component is individually
discoverable and self-contained, without changing any public API.

- Move SkeletonText and SkeletonCircle into self-contained modules under
  skeleton/derivatives/<name>/ (implementation, types, index, docs,
  stories). The base Skeleton stays at the component root and owns the
  shared nimbusSkeleton recipe; derivatives import it directly per the
  cross-chunk-import rule. Public @commercetools/nimbus exports are
  unchanged, re-exported through the skeleton barrel.
- Give Skeleton, SkeletonText, and SkeletonCircle each their own
  .mdx + .dev.mdx + .docs.spec.tsx so every component gets its own nav
  entry in the docs app (Components -> Feedback) and its own Storybook
  sidebar entry via split .stories.tsx files. The base skeleton docs are
  refocused onto the primitive and cross-link to the derivatives.
- Rename the OpenSpec capability skeleton -> nimbus-skeleton to match the
  nimbus-<name> convention; keep it a single family capability with the
  derivatives as labeled requirement blocks.

Note: this intentionally departs from the repo's one-directory-per-nav-entry
convention. Every other component directory has exactly one main .mdx; here
three main .mdx files (skeleton, skeleton-text, skeleton-circle) live under a
single skeleton/ tree, plus a nested derivatives/ folder that exists nowhere
else. It works because docs discovery is a recursive packages/**/*.mdx glob
with routing driven purely by frontmatter (id/menu/exportName) — file path and
directory don't gate a page, so each component surfaces as its own document.
The trade-off (no prior precedent) was accepted deliberately to make each
family member individually discoverable while keeping the code cohesive.
