+++ openspec/changes/add-skeleton-component/specs/nimbus-skeleton/spec.md
