+++ packages/nimbus/src/components/skeleton/derivatives/skeleton-circle/index.ts
@@ -1,0 +1,2 @@
+export { SkeletonCircle } from "./skeleton-circle";
+export type { SkeletonCircleProps } from "./skeleton-circle.types";
