+++ packages/nimbus/src/components/skeleton/derivatives/skeleton-text/index.ts
@@ -1,0 +1,2 @@
+export { SkeletonText } from "./skeleton-text";
+export type { SkeletonTextProps } from "./skeleton-text.types";
