+++ packages/nimbus/src/components/skeleton/index.ts
@@ -1,7 +1,9 @@
-export { Skeleton, SkeletonText, SkeletonCircle } from "./skeleton";
-export type {
-  SkeletonProps,
-  SkeletonTextProps,
-  SkeletonCircleProps,
-  SkeletonRecipeProps,
-} from "./skeleton.types";
+export { Skeleton } from "./skeleton";
+export type { SkeletonProps, SkeletonRecipeProps } from "./skeleton.types";
+
+// Derivatives — spin-offs built on the base Skeleton, each self-contained under
+// `derivatives/`.
+export { SkeletonText } from "./derivatives/skeleton-text";
+export type { SkeletonTextProps } from "./derivatives/skeleton-text";
+export { SkeletonCircle } from "./derivatives/skeleton-circle";
+export type { SkeletonCircleProps } from "./derivatives/skeleton-circle";
