+++ packages/nimbus/src/components/skeleton/skeleton.dev.mdx
@@ -9,19 +9,19 @@ tab-order: 3
 ### Import
 
 ```tsx
-import {
-  Skeleton,
-  SkeletonText,
-  SkeletonCircle,
-  type SkeletonProps,
-  type SkeletonTextProps,
-  type SkeletonCircleProps,
-} from '@commercetools/nimbus';
+import { Skeleton, type SkeletonProps } from '@commercetools/nimbus';
 ```
+
+For paragraph and circular placeholders, see
+[SkeletonText](/components/feedback/skeletontext) and
+[SkeletonCircle](/components/feedback/skeletoncircle) — both are built on
+`Skeleton`.
 
 ### Basic usage
 
-The Skeleton component renders a muted, animated placeholder to hold space for content while it loads. Use it to reduce perceived layout shift and communicate loading state visually.
+`Skeleton` renders a muted, animated block that holds space for content while it
+loads. Give it a `width` and `height` (or any Chakra style props) to match the
+content it stands in for.
 
 ```jsx live-dev
 const App = () => (
@@ -33,7 +33,10 @@ const App = () => (
 
 ### Conditional rendering (standalone pattern)
 
-Skeleton does **not** have an `isLoaded` prop. Instead, render the skeleton during loading and swap it for real content using conditional rendering. This keeps loading-state management in your own code, separate from the styling component.
+Skeleton does **not** have an `isLoaded` prop. Instead, render the skeleton
+during loading and swap it for real content using conditional rendering. This
+keeps loading-state management in your own code, separate from the styling
+component.
 
 ```jsx live-dev
 const App = () => {
@@ -96,71 +99,24 @@ const App = () => (
 )
 ```
 
-### Multi-line text skeleton
-
-Use `SkeletonText` to approximate a paragraph of text. The last line is narrower by default (`lastLineWidth="60%"`) to mimic a natural paragraph ending.
-
-```jsx live-dev
-const App = () => (
-  <SkeletonText lines={4} width="300px" lastLineWidth="40%" />
-)
-```
-
-### Matching text size with `textStyle`
-
-`SkeletonText` accepts a `textStyle` prop (default `"body"`) that makes the
-placeholder lines occupy the same vertical rhythm as real text of that style. It
-sets the container's `font-size` and `line-height`, so each bar's height (`1em`)
-and the inter-line gap (`calc(1lh - 1em)`, the style's leading) scale
-automatically — a `body` skeleton uses a 26 px pitch, a `3xl` heading skeleton a
-36 px pitch. Pass any Nimbus text style to match the content you are standing in
-for.
-
-```jsx live-dev
-const App = () => (
-  <Stack gap="600" width="300px">
-    <SkeletonText textStyle="caption" />
-    <SkeletonText textStyle="body" />
-    <SkeletonText textStyle="3xl" lines={2} />
-  </Stack>
-)
-```
-
-You can still override the derived sizing with the `lineHeight` (bar height) and
-`spacing` (gap) props when you need a bespoke layout.
-
-### Circle skeleton
-
-`SkeletonCircle` is a convenience wrapper over `Skeleton shape="circle"`. Use
-the avatar-aligned `size` prop (`"2xs"` / `"xs"` / `"md"`) to stand in for an
-`Avatar`, or `boxSize` for a custom dimension. With neither set it defaults to a
-`1em` circle (handy inline next to text).
-
-```jsx live-dev
-const App = () => (
-  <Stack direction="row" gap="400" alignItems="center">
-    <SkeletonCircle />
-    <SkeletonCircle size="2xs" />
-    <SkeletonCircle size="xs" />
-    <SkeletonCircle size="md" />
-    <SkeletonCircle boxSize="4800" />
-  </Stack>
-)
-```
-
 ## Component requirements
 
-## Accessibility
+### Accessibility
 
-Skeleton placeholders are **decorative by default** — all three components (`Skeleton`, `SkeletonText`, `SkeletonCircle`) render with `aria-hidden="true"` so screen readers skip over them. This is the correct behavior: the shapes convey visual loading activity but carry no semantic meaning on their own.
+Skeleton placeholders are **decorative by default** — they render with
+`aria-hidden="true"` so screen readers skip over them. This is the correct
+behavior: the shapes convey visual loading activity but carry no semantic
+meaning on their own.
 
-To communicate loading state to assistive technology, set `aria-busy="true"` on the **container** that wraps the skeleton region. When content has loaded, remove the attribute (or set it to `false`) and render the real content in its place.
+To communicate loading state to assistive technology, set `aria-busy="true"` on
+the **container** that wraps the skeleton region. When content has loaded, remove
+the attribute (or set it to `false`) and render the real content in its place.
 
 ```tsx
 // Announce the loading region to screen readers via aria-busy
 <section aria-busy={isLoading} aria-label="Product details">
   {isLoading ? (
-    <SkeletonText lines={3} width="100%" />
+    <Skeleton width="100%" height="120px" />
   ) : (
     <ProductDetails product={product} />
   )}
@@ -173,45 +129,41 @@ un-hiding it (`aria-hidden={false}`) is an invalid ARIA combination that
 accessibility linters (axe) flag as `aria-prohibited-attr`. Announce the loading
 region on the wrapping container instead.
 
-### Reduced motion
+#### Reduced motion
 
-When the user has enabled the `prefers-reduced-motion: reduce` media condition, all animations are automatically disabled — regardless of the `animation` prop value. The `::after` wave sweep is also hidden. No additional configuration is required.
+When the user has enabled the `prefers-reduced-motion: reduce` media condition,
+all animations are automatically disabled — regardless of the `animation` prop
+value. The `::after` wave sweep is also hidden. No additional configuration is
+required.
 
-If your use case requires tracking and analytics for this component, it is good practice to add a **persistent**, **unique** id to the component:
+If your use case requires tracking and analytics for this component, it is good
+practice to add a **persistent**, **unique** id to the component:
 
 ```tsx
 const PERSISTENT_ID = "product-detail-skeleton";
 
 export const ProductDetailSkeleton = () => (
   <div id={PERSISTENT_ID} aria-busy>
-    <SkeletonText lines={4} width="100%" />
+    <Skeleton width="100%" height="120px" />
   </div>
 );
 ```
 
 #### Keyboard navigation
 
-Skeleton components are non-interactive and decorative. They do not receive focus or support keyboard interaction.
+Skeleton components are non-interactive and decorative. They do not receive
+focus or support keyboard interaction.
 
 ## API reference
 
-### Skeleton
-
 <PropsTable id="Skeleton" />
-
-### SkeletonText
-
-<PropsTable id="SkeletonText" />
-
-### SkeletonCircle
-
-<PropsTable id="SkeletonCircle" />
 
 ## Common patterns
 
 ### Composing a card skeleton
 
-Combine the three components to build realistic loading layouts.
+Combine `Skeleton` with `SkeletonCircle` and `SkeletonText` to build realistic
+loading layouts that mirror the final content.
 
 ```jsx live-dev
 const App = () => (
@@ -234,3 +186,5 @@ These examples demonstrate how to test your implementation when using Skeleton w
 ## Resources
 
 - [Storybook](https://nimbus-storybook.vercel.app/?path=/docs/components-skeleton--docs)
+- [SkeletonText](/components/feedback/skeletontext)
+- [SkeletonCircle](/components/feedback/skeletoncircle)
