+++ packages/nimbus/src/components/skeleton/skeleton.types.ts
@@ -64,91 +64,3 @@ export type SkeletonProps = OmitInternalProps<SkeletonRootSlotProps> & {
    */
   [key: `data-${string}`]: unknown;
 };
-
-/**
- * Props for the SkeletonText component.
- * Renders a stack of N placeholder lines to approximate a paragraph of text.
- *
- * Note: `shape` is intentionally absent — SkeletonText lines are always
- * rectangles. Use `Skeleton shape="circle"` or `SkeletonCircle` for circular
- * placeholders.
- */
-export type SkeletonTextProps = OmitInternalProps<HTMLChakraProps<"div">> & {
-  /**
-   * Number of placeholder lines to render.
-   * @default 3
-   */
-  lines?: number;
-  /**
-   * Nimbus text style the placeholder should visually match (e.g. `"body"`,
-   * `"caption"`, `"3xl"`). Sets the container's `font-size` and `line-height`,
-   * so the default bar height (`1em`) and line spacing (`calc(1lh - 1em)`) scale
-   * to give the same vertical rhythm as real text of that style.
-   * @default "body"
-   */
-  textStyle?: HTMLChakraProps<"div">["textStyle"];
-  /**
-   * Height of each placeholder line. Overrides the `textStyle`-derived height.
-   * @default "1em"
-   */
-  lineHeight?: string | number;
-  /**
-   * Gap between lines (accepts Nimbus spacing tokens or CSS values). Defaults to
-   * the `textStyle`'s leading (`line-height − font-size`) via the `1lh`/`1em`
-   * units; pass an explicit value to override.
-   * @default "calc(1lh - 1em)"
-   */
-  spacing?: string | number;
-  /**
-   * Width of the last line, narrower to mimic a paragraph ending.
-   * @default "60%"
-   */
-  lastLineWidth?: string | number;
-  /**
-   * Animation forwarded to each Skeleton line.
-   * @default "pulse"
-   */
-  animation?: SkeletonRecipeProps["animation"];
-  /**
-   * Whether the container is hidden from assistive technology.
-   * @default true
-   */
-  "aria-hidden"?: boolean | "true" | "false";
-  /**
-   * Ref forwarding to the root container element.
-   */
-  ref?: React.Ref<HTMLDivElement>;
-};
-
-/**
- * Props for the SkeletonCircle component.
- * Renders a circular loading placeholder. Use the avatar-aligned `size` prop to
- * match an Avatar, or `boxSize` for a custom dimension — SkeletonCircle applies
- * whichever you set to equal width and height. `width`/`height` are omitted in
- * favor of these.
- */
-export type SkeletonCircleProps = OmitInternalProps<
-  HTMLChakraProps<"div">,
-  "width" | "height"
-> & {
-  /**
-   * Named circle size, aligned with the `Avatar` size scale
-   * (`2xs` = 24px, `xs` = 32px, `md` = 40px). For a custom dimension use
-   * `boxSize` instead. When neither is set, the circle defaults to `1em`.
-   */
-  size?: "2xs" | "xs" | "md";
-  /**
-   * Animation of the skeleton placeholder.
-   * @default "pulse"
-   */
-  animation?: SkeletonRecipeProps["animation"];
-  /**
-   * Whether the element is hidden from assistive technology.
-   * @default true
-   */
-  "aria-hidden"?: boolean | "true" | "false";
-  /**
-   * Ref forwarding to the root element.
-   */
-  ref?: React.Ref<HTMLDivElement>;
-};
